+++ sopa/aggregation/transcripts.py
@@ -23,6 +23,7 @@ def count_transcripts(
     shapes_key: str | None = None,
     points_key: str | None = None,
     geo_df: gpd.GeoDataFrame | None = None,
+    only_excluded: bool = False,
 ) -> AnnData:
     """Counts transcripts per cell.
 
@@ -32,6 +33,7 @@ def count_transcripts(
         shapes_key: Key of `sdata` containing the cell boundaries. If only one `shapes` element, this does not have to be provided.
         points_key: Key of `sdata` containing the transcripts. If only one `points` element, this does not have to be provided.
         geo_df: If the cell boundaries are not yet in `sdata`, a `GeoDataFrame` can be directly provided for cell boundaries
+        only_excluded: By default, the genes matching the pattern in `sopa.settings.gene_exclude_pattern` are excluded from the count. If `only_excluded=True`, it counts **only** these excluded genes.
 
     Returns:
         An `AnnData` object of shape `(n_cells, n_genes)` with the counts per cell
@@ -47,16 +49,22 @@ def count_transcripts(
     gene_column = gene_column or get_feature_key(points, raise_error=True)
 
     log.info(f"Aggregating transcripts over {len(geo_df)} cells")
-    return _count_transcripts_aligned(geo_df, points, gene_column)
+    return _count_transcripts_aligned(geo_df, points, gene_column, only_excluded)
 
 
-def _count_transcripts_aligned(geo_df: gpd.GeoDataFrame, points: dd.DataFrame, value_key: str) -> AnnData:
+def _count_transcripts_aligned(
+    geo_df: gpd.GeoDataFrame,
+    points: dd.DataFrame,
+    value_key: str,
+    only_excluded: bool = False,
+) -> AnnData:
     """Count transcripts per cell. The cells and points have to be aligned (i.e., in the same coordinate system)
 
     Args:
         geo_df: Cells geometries
         points: Transcripts dataframe
         value_key: Key of `points` containing the genes names
+        only_excluded: Whether to count the valid genes or the excluded ones
 
     Returns:
         An `AnnData` object of shape `(n_cells, n_genes)` with the counts per cell
@@ -74,7 +82,14 @@ def _count_transcripts_aligned(geo_df: gpd.GeoDataFrame, points: dd.DataFrame, v
 
     with ProgressBar():
         points.map_partitions(
-            partial(_add_csr, X_partitions, geo_df, gene_column=value_key, gene_names=gene_names),
+            partial(
+                _add_csr,
+                X_partitions,
+                geo_df,
+                gene_column=value_key,
+                gene_names=gene_names,
+                only_excluded=only_excluded,
+            ),
             meta=(),
         ).compute()
 
@@ -82,7 +97,8 @@ def _count_transcripts_aligned(geo_df: gpd.GeoDataFrame, points: dd.DataFrame, v
         adata.X += X_partition
 
     if settings.gene_exclude_pattern is not None:
-        adata = adata[:, ~adata.var_names.str.match(settings.gene_exclude_pattern, case=False, na=False)].copy()
+        matching = adata.var_names.str.match(settings.gene_exclude_pattern, case=False, na=True)
+        adata = (adata[:, matching] if only_excluded else adata[:, ~matching]).copy()
 
     return adata
 
@@ -93,9 +109,11 @@ def _add_csr(
     partition: pd.DataFrame,
     gene_column: str,
     gene_names: list[str],
+    only_excluded: bool,
 ) -> None:
     if settings.gene_exclude_pattern is not None:
-        partition = partition[~partition[gene_column].str.match(settings.gene_exclude_pattern, case=False, na=False)]
+        matching = partition[gene_column].str.match(settings.gene_exclude_pattern, case=False, na=True)
+        partition = partition[matching] if only_excluded else partition[~matching]
 
     points_gdf = gpd.GeoDataFrame(partition, geometry=gpd.points_from_xy(partition["x"], partition["y"]))
     joined = geo_df.sjoin(points_gdf)